Refactor App Structure and Introduce Routing for Enhanced Navigation
- Reorganized the main application structure by separating the main app page and form flow page into distinct components.
- Implemented React Router for improved navigation between the main app and form flow views.
- Updated the MainLayout and ResponsiveNavbar to include navigation options for the new form flow feature.
- Removed unused imports and streamlined the app.tsx file for better readability and maintainability.
