+++ apps/prompttoform/src/app/components/templates/ResponsiveNavbar.tsx
@@ -4,12 +4,16 @@ interface ResponsiveNavbarProps {
   onSettingsClick: () => void;
   onHistoryClick: () => void;
   onImportJsonClick: () => void;
+  onFormFlowClick?: () => void;
+  showFormFlowButton?: boolean;
 }
 
 export function ResponsiveNavbar({
   onSettingsClick,
   onHistoryClick,
   onImportJsonClick,
+  onFormFlowClick,
+  showFormFlowButton = false,
 }: ResponsiveNavbarProps) {
   const [isMobileMenuOpen, setIsMobileMenuOpen] = useState(false);
 
@@ -39,6 +43,27 @@ export function ResponsiveNavbar({
           {/* Desktop Navigation */}
           <div className="hidden md:block">
             <div className="ml-10 flex items-baseline space-x-4">
+              {showFormFlowButton && onFormFlowClick && (
+                <button
+                  onClick={onFormFlowClick}
+                  className="inline-flex items-center px-3 py-2 border border-transparent shadow-sm text-sm font-medium rounded-md text-white bg-indigo-600 hover:bg-indigo-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500 transition-colors"
+                >
+                  <svg
+                    className="w-3 h-3 mr-2"
+                    fill="none"
+                    stroke="currentColor"
+                    viewBox="0 0 24 24"
+                  >
+                    <path
+                      strokeLinecap="round"
+                      strokeLinejoin="round"
+                      strokeWidth={2}
+                      d="M13 10V3L4 14h7v7l9-11h-7z"
+                    />
+                  </svg>
+                  View Flow
+                </button>
+              )}
               <button
                 onClick={onImportJsonClick}
                 className="inline-flex items-center px-3 py-2 border border-zinc-300 shadow-sm text-sm font-medium rounded-md text-zinc-700 bg-white hover:bg-zinc-50 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500 transition-colors"
@@ -122,6 +147,32 @@ export function ResponsiveNavbar({
       {/* Mobile menu */}
       <div className={`${isMobileMenuOpen ? 'block' : 'hidden'} md:hidden`}>
         <div className="px-2 pt-2 pb-3 space-y-1 sm:px-3 bg-white border-t border-zinc-200">
+          {showFormFlowButton && onFormFlowClick && (
+            <button
+              onClick={() => {
+                onFormFlowClick();
+                closeMobileMenu();
+              }}
+              className="w-full text-left block px-3 py-2 rounded-md text-base font-medium text-white bg-indigo-600 hover:bg-indigo-700 focus:outline-none focus:ring-2 focus:ring-inset focus:ring-indigo-500 transition-colors"
+            >
+              <div className="flex items-center">
+                <svg
+                  className="w-3 h-3 mr-2"
+                  fill="none"
+                  stroke="currentColor"
+                  viewBox="0 0 24 24"
+                >
+                  <path
+                    strokeLinecap="round"
+                    strokeLinejoin="round"
+                    strokeWidth={2}
+                    d="M13 10V3L4 14h7v7l9-11h-7z"
+                  />
+                </svg>
+                View Flow
+              </div>
+            </button>
+          )}
           <button
             onClick={() => {
               onImportJsonClick();
